requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.13.1
elasticsearch==6.4.0
Flask==3.0.3
flask-restx==1.3.0
frozenlist==1.4.1
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
MarkupSafe==2.1.5
multidict==6.0.5
pip==22.0.2
pyTelegramBotAPI==4.19.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==59.6.0
urllib3==1.26.18
Werkzeug==3.0.3
yarl==1.9.4